+++ src/app/components/LoginForm/Form/Form.tsx
@@ -4,11 +4,10 @@ import Link from "next/link";
 import { useActionState } from "react";
 
 import { Button } from "@/app/components/ui/button";
-import { Input } from "@/app/components/ui/input";
-import { Label } from "@/app/components/ui/label";
 import type { LoginFormState } from "@/app/components/LoginForm/action";
 import { Asterisk } from "lucide-react";
 import ErrorCard from "@/app/components/ErrorCard/ErrorCard";
+import FormInputField from "../../FormInputField/FormInputField";
 
 /**
  * フォームの初期状態
@@ -40,25 +39,26 @@ const Form = ({
     <form action={formAction} className="space-y-4">
       <ErrorCard errors={state.errors} />
       <div className="space-y-2">
-        <Label htmlFor="email" className="gap-0.5">
-          <Asterisk />
-          メールアドレス
-        </Label>
-        <Input
+        <FormInputField
           id="email"
           name="email"
           type="email"
           placeholder="name@example.com"
           required
           autoComplete="email"
+          label={"メールアドレス"}
         />
       </div>
       <div className="space-y-2">
+        <FormInputField
+          id="password"
+          name="password"
+          type="password"
+          required
+          autoComplete="current-password"
+          label={"パスワード"}
+        />
         <div className="flex items-center justify-between">
-          <Label htmlFor="password" className="gap-0.5">
-            <Asterisk />
-            パスワード
-          </Label>
           <Link
             href="/forgot-password"
             className="text-sm text-primary hover:underline"
@@ -66,13 +66,6 @@ const Form = ({
             パスワードをお忘れですか？
           </Link>
         </div>
-        <Input
-          id="password"
-          name="password"
-          type="password"
-          required
-          autoComplete="current-password"
-        />
       </div>
       <Button type="submit" className="w-full" disabled={isPending}>
         {isPending ? "送信中..." : "ログイン"}
