+++ src/app/components/SignUpForm/Form/Form.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { Input } from "@/app/components/ui/input";
-import { Label } from "@/app/components/ui/label";
 import { Button } from "@/app/components/ui/button";
 import { X } from "lucide-react";
 import { useActionState, useMemo, useState } from "react";
 import type { SignUpFormState } from "../action";
 import { passwordMatchSchema } from "@/app/zod/scheme";
 import ErrorCard from "@/app/components/ErrorCard/ErrorCard";
+import FormInputField from "../../FormInputField/FormInputField";
 
 const initialState: SignUpFormState = {
   errors: undefined,
@@ -48,29 +47,28 @@ const Form = ({
     <form action={formAction} className="space-y-4">
       <ErrorCard errors={state.errors} />
       <div className="space-y-2">
-        <Label htmlFor="username">ユーザー名</Label>
-        <Input
+        <FormInputField
           id="username"
           name="username"
           placeholder="johndoe"
           required
           autoComplete="username"
+          label={"ユーザー名"}
         />
       </div>
       <div className="space-y-2">
-        <Label htmlFor="email">メールアドレス</Label>
-        <Input
+        <FormInputField
           id="email"
           name="email"
           type="email"
           placeholder="name@example.com"
           required
           autoComplete="email"
+          label={"メールアドレス"}
         />
       </div>
       <div className="space-y-2">
-        <Label htmlFor="password">パスワード</Label>
-        <Input
+        <FormInputField
           id="password"
           name="password"
           type="password"
@@ -78,11 +76,11 @@ const Form = ({
           onChange={(e) => setPassword(e.target.value)}
           required
           autoComplete="new-password"
+          label={"パスワード"}
         />
       </div>
       <div className="space-y-2">
-        <Label htmlFor="passwordConfirm">パスワード（確認）</Label>
-        <Input
+        <FormInputField
           id="passwordConfirm"
           name="passwordConfirm"
           type="password"
@@ -90,6 +88,7 @@ const Form = ({
           onChange={(e) => setPasswordConfirm(e.target.value)}
           required
           autoComplete="new-password"
+          label={"パスワード（確認）"}
         />
 
         {passwordValidationErrors?.passwordConfirm?.map((error) => (
